LPD-96669 Rename the depot scopes tests to asset library scopes
